comment delete opertion upadted,deleted comment id removal done

+++ backend/src/controllers/comment.controller.js
@@ -1,32 +1,32 @@
 import { Comment } from "../models/comment.model.js";
-import {Post} from "../models/post.model.js";
+import { Post } from "../models/post.model.js";
 
 const createComment = async (req, res, next) => {
     try {
-        const {ownerId, text, content, postId} = req.body;
-        if(!ownerId || !postId){
-            res.status(400).json({message: "owner _id and post _id is required"})
+        const { ownerId, text, content, postId } = req.body;
+        if (!ownerId || !postId) {
+            res.status(400).json({ message: "owner _id and post _id is required" })
         }
 
-        if(!text){
-            res.status(400).json({message: "comment must at least contain text"});
+        if (!text) {
+            res.status(400).json({ message: "comment must at least contain text" });
         }
 
         await Comment.create(req.body)
-        .then(result=>{
-            Post.findByIdAndUpdate(result.postId, {$push: {comments: result._id}}, {new: true})
-            .exec()
-            .then(post=>{
-                if(!post){
-                    res.status(400).json({message: "post id incorrect or not registred"});
-                }
-                res.status(200).json({message: "comment created succefuly", comment: result})
+            .then(result => {
+                Post.findByIdAndUpdate(result.postId, { $push: { comments: result._id } }, { new: true })
+                    .exec()
+                    .then(post => {
+                        if (!post) {
+                            res.status(400).json({ message: "post id incorrect or not registred" });
+                        }
+                        res.status(200).json({ message: "comment created succefuly", comment: result })
+                    })
+                // res.status(200).json({message: "comment created succefuly", detail: result})
             })
-            // res.status(200).json({message: "comment created succefuly", detail: result})
-        })
 
     } catch (error) {
-        res.status(500).json({message: "Internal server error", error: error.message});
+        res.status(500).json({ message: "Internal server error", error: error.message });
     }
 };
 
@@ -34,48 +34,60 @@ const createComment = async (req, res, next) => {
 
 const getComment = async (req, res, next) => {
     try {
-        await Comment.find({ownerId: req.params.id})
-        .exec()
-        .then(result=>{
-            if(!result){
-                res.status(400).json({message: "owner id not registred or incorrect"});
-            }
+        await Comment.find({ ownerId: req.params.id })
+            .exec()
+            .then(result => {
+                if (!result) {
+                    res.status(400).json({ message: "owner id not registred or incorrect" });
+                }
 
-            res.status(200).json({message: "user comments", comments: result});
-        })
+                res.status(200).json({ message: "user comments", comments: result });
+            })
     } catch (error) {
-        res.status(500).json({message: "Internal server error", error: error.message});
+        res.status(500).json({ message: "Internal server error", error: error.message });
     }
 };
 
 const editComment = async (req, res, next) => {
-    const {text} = req.body
+    const { text } = req.body
     try {
-        await Comment.findOneAndUpdate({ownerId: req.params.id}, {text}, {new: true})
-        .exec()
-        .then(result=>{
-            if(!result){
-                res.status(400).json({message: "owner id incorrect or not registred"})
-            }
-            res.status(200).json({message: "comment edited successfuly"});
-        })
+        await Comment.findByIdAndUpdate(req.params.id, { text }, { new: true })
+            .exec()
+            .then(result => {
+                if (!result) {
+                    res.status(400).json({ message: "comment id incorrect or not registred", state: false })
+                }
+                res.status(200).json({ message: "comment edited successfuly", state: true });
+            })
     } catch (error) {
-        res.status(500).json({message: "Internal server error", error: error.message})
+        res.status(500).json({ message: "Internal server error", error: error.message })
     }
 };
 
 const deleteComment = async (req, res, next) => {
     try {
-        await Comment.findByIdAndDelete(req.params.id)
-        .exec()
-        .then(result=>{
-            if(!result){
-                res.status(400).json({message: "comment id incorrect or not registred"})
-            }
-            res.status(200).json({message: "comment deleted successfuly"})
-        })
+        const { postId } = req.body;
+        await Post.findByIdAndUpdate(postId, { $pull: { comments: req.params.id } }, { new: true })
+            .exec()
+            .then(resp => {
+                if (!resp) {
+                    res.json({ message: "incorrect post id comment not found " })
+                }
+                if (resp) {
+                    Comment.findByIdAndDelete(req.params.id)
+                        .exec()
+                        .then(result => {
+                            if (!result) {
+                                res.status(400).json({ message: "comment id incorrect or not registred", state: true })
+                            }
+                            res.status(200).json({ message: "comment deleted successfuly", state: false })
+                        })
+                }
+
+            })
+
     } catch (error) {
-        res.status(500).json({message: "Internal server error", error: error.message})
+        res.status(500).json({ message: "Internal server error", error: error.message })
     }
 };
